+++ assets/javascripts/kitten/components/modals/modal.js
@@ -1,4 +1,5 @@
 import React, { Component } from 'react'
+import PropTypes from 'prop-types'
 import classNames from 'classnames'
 // Via "https://github.com/reactjs/react-modal"
 import ReactModal from 'react-modal'
@@ -52,6 +53,8 @@ export class Modal extends Component {
       trigger,
       content,
       label,
+      labelledby,
+      describedby,
       className,
       onClose,
       ...others
@@ -64,6 +67,7 @@ export class Modal extends Component {
         {this.renderTriggerAction()}
 
         <ReactModal
+          role="dialog"
           className={{
             base: 'k-Modal__content',
             afterOpen: 'k-Modal--afterOpen',
@@ -75,6 +79,10 @@ export class Modal extends Component {
             beforeClose: 'k-Modal__overlay--beforeClose',
           }}
           isOpen={this.state.showModal}
+          aria={{
+            labelledby,
+            describedby,
+          }}
           onRequestClose={this.close}
           contentLabel={label}
         >
@@ -87,6 +95,14 @@ export class Modal extends Component {
   }
 }
 
+Modal.propTypes = {
+  label: PropTypes.string,
+  labelledby: PropTypes.string,
+  describedby: PropTypes.string,
+}
+
 Modal.defaultProps = {
   label: 'Modal',
+  labelledby: '',
+  describedby: '',
 }