+++ CHANGELOG.md
@@ -4,6 +4,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
+- Feature: Update A11Y on `Modal` component.
+
 ## [20.7.1] - 2018-04-11
 
 Fix: